feat(sdk): add more computer use action dto and fix unmarshal bug
- Add MouseDoubleClickAction, MouseMoveAction, MouseScrollAction, MouseDragAction, KeyboardTypeAction, KeyboardHotkeyAction, ScreenshotAction, WaitAction, FinishedAction, FailedAction DTOs.
- Add a helper function to unmarshal Length fields to fix a bug where the entire JSON source was being used for deserialization instead of the sub-object.
- Refactor all UnmarshalJSON methods to use the new helper function.
